Just tons of stuff. Detail screen is more useful

+++ app/src/main/java/com/yoxjames/openstitch/navigation/NavigationState.kt
@@ -7,7 +7,7 @@ object None : NavigationScreenState
 object PatternList : NavigationScreenState
 
 data class PatternDetail(
-    val patternId: Int,
+    val patternId: Long,
 ) : NavigationScreenState
 
 @JvmInline
